画像パスがある場合のみフォルダを削除するようコード作成

+++ src/main/java/com/study/skillup/restfulapi/service/ProductService.java
@@ -57,21 +57,23 @@ public class ProductService {
   }
 
   public void delete(long id) throws IOException {
+
     Product product = productRepository.findById(id);
     String imageFIlePath = product.getImage_path();
-    Path filePath =
-        Paths.get(
-            "/Users/yuusuke/study/skillup/yusuke-restful-api/src/main/resources/static/images/",
-            "商品" + id,
-            imageFIlePath);
-    Files.delete(filePath);
+    if (imageFIlePath != null) {
+      Path filePath =
+          Paths.get(
+              "/Users/yuusuke/study/skillup/yusuke-restful-api/src/main/resources/static/images/",
+              "商品" + id,
+              imageFIlePath);
+      Files.delete(filePath);
 
-    Path folderPath =
-        Paths.get(
-            "/Users/yuusuke/study/skillup/yusuke-restful-api/src/main/resources/static/images/",
-            "商品" + id);
-    Files.deleteIfExists(folderPath);
-
+      Path folderPath =
+          Paths.get(
+              "/Users/yuusuke/study/skillup/yusuke-restful-api/src/main/resources/static/images/",
+              "商品" + id);
+      Files.deleteIfExists(folderPath);
+    }
     Product deleted_product = productRepository.findById(id);
     productRepository.delete(deleted_product);
   }
